Added benchmark based on pytest-benchmark.
