+++ doc/requirements.txt
@@ -1,7 +1,7 @@
 -r ../requirements.txt
 
 # Enforce latest version on ReadTheDocs
-sphinx>=4.4.0
+sphinx>=4.5.0
 pyTooling>=1.9.5
 
 # Sphinx Extenstions
@@ -13,7 +13,7 @@ sphinxcontrib-mermaid>=0.7.1
 #sphinxcontrib-spelling>=2.2.0
 autoapi
 sphinx_fontawesome>=0.0.6
-sphinx_autodoc_typehints>=1.17.0
+sphinx_autodoc_typehints>=1.18.1
 # changelog>=0.3.5
 
 # BuildTheDocs Extensions (mostly patched Sphinx extensions)
