+++ tests/benchmark/MetaClasses/__init__.py
@@ -28,133 +28,4 @@
 # SPDX-License-Identifier: Apache-2.0                                                                                  #
 # ==================================================================================================================== #
 #
-"""Performance tests for SlottedType."""
-from pyTooling.MetaClasses import SlottedType
-from . import PerformanceTest
-
-
-if __name__ == "__main__":  # pragma: no cover
-	print("ERROR: you called a testcase declaration file as an executable module.")
-	print("Use: 'python -m unitest <testcase module>'")
-	exit(1)
-
-
-class Node_1:
-	_data_0: int
-
-	def __init__(self, data):
-		self._data_0 = data
-
-	def inc(self, add: int):
-		self._data_0 = self._data_0 + add
-
-
-class SlottedNode_1(metaclass=SlottedType):
-	_data_0: int
-
-	def __init__(self, data):
-		self._data_0 = data
-
-	def inc(self, add: int):
-		self._data_0 = self._data_0 + add
-
-
-class Node_10:
-	_data_0: int
-	_data_1: int
-	_data_2: int
-	_data_3: int
-	_data_4: int
-	_data_5: int
-	_data_6: int
-	_data_7: int
-	_data_8: int
-	_data_9: int
-
-	def __init__(self, data):
-		self._data_0 = data
-		self._data_1 = data
-		self._data_2 = data
-		self._data_3 = data
-		self._data_4 = data
-		self._data_5 = data
-		self._data_6 = data
-		self._data_7 = data
-		self._data_8 = data
-		self._data_9 = data
-
-
-class SlottedNode_10(metaclass=SlottedType):
-	_data_0: int
-	_data_1: int
-	_data_2: int
-	_data_3: int
-	_data_4: int
-	_data_5: int
-	_data_6: int
-	_data_7: int
-	_data_8: int
-	_data_9: int
-
-	def __init__(self, data):
-		self._data_0 = data
-		self._data_1 = data
-		self._data_2 = data
-		self._data_3 = data
-		self._data_4 = data
-		self._data_5 = data
-		self._data_6 = data
-		self._data_7 = data
-		self._data_8 = data
-		self._data_9 = data
-
-
-class Tree(PerformanceTest):
-	def test_CreateObjects_1(self):
-		def wrapperNormal(count: int):
-			def func():
-				nodes = [Node_1(i) for i in range(count)]
-
-			return func
-
-		def wrapperSlotted(count: int):
-			def func():
-				nodes = [SlottedNode_1(i) for i in range(count)]
-
-			return func
-
-		self.runTests(wrapperNormal, wrapperSlotted, self.counts)
-
-	def test_CreateObjects_10(self):
-		def wrapperNormal(count: int):
-			def func():
-				nodes = [Node_10(i) for i in range(count)]
-
-			return func
-
-		def wrapperSlotted(count: int):
-			def func():
-				nodes = [SlottedNode_10(i) for i in range(count)]
-
-			return func
-
-		self.runTests(wrapperNormal, wrapperSlotted, self.counts)
-
-	def test_Accumulate_1(self):
-		def wrapperNormal(count: int):
-			def func():
-				node = Node_1(1)
-				for i in range(count):
-					node.inc(1)
-
-			return func
-
-		def wrapperSlotted(count: int):
-			def func():
-				node = SlottedNode_1(1)
-				for i in range(count):
-					node.inc(1)
-
-			return func
-
-		self.runTests(wrapperNormal, wrapperSlotted, self.counts)
+"""Performance tests for MetaClasses."""
