+++ tests/unit/MetaClasses/SlottedType.py
